requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
args==0.1.0
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
clint==0.5.1
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
gspread==4.0.1
html-testRunner==1.2.1
httplib2==0.19.1
idna==3.2
iniconfig==1.1.1
Jinja2==2.11.3
lxml==4.6.4
MarkupSafe==1.1.1
mysql-connector-python==8.0.28
oauth2client==4.1.3
oauthlib==3.1.1
packaging==20.9
pluggy==0.13.1
protobuf==3.19.4
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.4
pytest-html==3.1.1
pytest-metadata==1.11.0
python-multipart==0.0.5
python-xml-hl7==1.3.0
rauth==0.7.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==3.141.0
six==1.11.0
soupsieve==2.3.1
toml==0.10.2
types-aiofiles==0.8.3
types-all==1.0.0
types-annoy==1.17.5
types-atomicwrites==1.4.3
types-backports==0.1.3
types-backports-abc==0.5.2
types-bleach==4.1.4
types-boto==2.49.7
types-cachetools==4.2.9
types-certifi==2021.10.8.1
types-characteristic==14.3.3
types-chardet==4.0.3
types-click==7.1.8
types-click-spinner==0.1.10
types-colorama==0.4.8
types-contextvars==2.4.2
types-croniter==1.0.7
types-cryptography==3.3.15
types-dataclasses==0.6.4
types-dateparser==1.0.14
types-DateTimeRange==1.2.3
types-decorator==5.1.4
types-Deprecated==1.2.5
types-docopt==0.6.5
types-docutils==0.17.5
types-emoji==1.2.7
types-enum34==1.1.8
types-fb303==1.0.0
types-filelock==3.2.5
types-first==2.0.2
types-Flask==1.1.6
types-freezegun==1.1.6
types-frozendict==2.0.5
types-futures==3.3.8
types-geoip2==3.0.0
types-ipaddress==1.0.8
types-itsdangerous==1.1.6
types-JACK-Client==0.5.4
types-Jinja2==2.11.9
types-kazoo==0.1.3
types-Markdown==3.3.12
types-MarkupSafe==1.1.10
types-maxminddb==1.5.0
types-mock==4.0.10
types-mypy-extensions==0.4.14
types-nmap==0.1.6
types-openssl-python==0.1.3
types-orjson==3.6.2
types-paramiko==2.8.13
types-pathlib2==2.3.0
types-Pillow==9.0.6
types-pkg-resources==0.1.3
types-polib==1.1.10
types-protobuf==3.19.12
types-pyaudio==0.2.9
types-pycurl==7.44.6
types-pyfarmhash==0.2.3
types-PyJWT==1.7.1
types-pymssql==2.1.0
types-PyMySQL==1.0.13
types-pyRFC3339==1.1.1
types-pysftp==0.2.9
types-python-dateutil==2.8.9
types-python-gflags==3.1.4
types-python-slugify==5.0.3
types-pytz==2021.3.5
types-pyvmomi==7.0.4
types-PyYAML==6.0.4
types-redis==4.1.16
types-requests==2.27.10
types-retry==0.9.5
types-Routes==2.5.0
types-scribe==2.0.0
types-simplejson==3.17.3
types-singledispatch==3.7.3
types-six==1.16.10
types-tabulate==0.8.5
types-termcolor==1.1.3
types-toml==0.10.4
types-tornado==5.1.1
types-typed-ast==1.5.2
types-tzlocal==4.1.0
types-ujson==4.2.1
types-urllib3==1.26.9
types-waitress==2.0.6
types-Werkzeug==1.0.9
types-xxhash==2.0.3
urllib3==1.26.4
webdriver-manager==3.5.1
xmltodict==0.12.0